# Improve RBF performance