--- a/docs/source/_summaries/pygem.radial.RBF._distance_matrix.rst
+++ b/docs/source/_summaries/pygem.radial.RBF._distance_matrix.rst
@@ -1,6 +1,0 @@
-pygem.radial.RBF._distance_matrix
-=================================
-
-.. currentmodule:: pygem.radial
-
-.. automethod:: RBF._distance_matrix