Improve performance for RBF `perform`.

Below few benchmarks to compare the old version and the new one (262144 mesh points, 8 control points):

| Old implementation | New Implementation |
| --- | --- |
| 18.7222251892 s | 0.127142906189 s |
| 18.5255019665 s | 0.118921041489 s |